+++ README.md
@@ -1,28 +1,19 @@
-# Arduin Uno Car APP
+# Tauri + Vue + TypeScript
 
-## 設定
+This template should help get you started developing with Vue 3 and TypeScript in Vite. The template uses Vue 3 `<script setup>` SFCs, check out the [script setup docs](https://v3.vuejs.org/api/sfc-script-setup.html#sfc-script-setup) to learn more.
 
-1. 鮑率預設為 `115200`
+## Recommended IDE Setup
 
-## 指令
+- [VS Code](https://code.visualstudio.com/) + [Volar](https://marketplace.visualstudio.com/items?itemName=Vue.volar) + [Tauri](https://marketplace.visualstudio.com/items?itemName=tauri-apps.tauri-vscode) + [rust-analyzer](https://marketplace.visualstudio.com/items?itemName=rust-lang.rust-analyzer)
 
-1. 顯示警告訊息
+## Type Support For `.vue` Imports in TS
 
-  > alert:\<訊息>
+Since TypeScript cannot handle type information for `.vue` imports, they are shimmed to be a generic Vue component type by default. In most cases this is fine if you don't really care about component prop types outside of templates. However, if you wish to get actual prop types in `.vue` imports (for example to get props validation when using manual `h(...)` calls), you can enable Volar's Take Over mode by following these steps:
 
-2. 播放自訂音效
+1. Run `Extensions: Show Built-in Extensions` from VS Code's command palette, look for `TypeScript and JavaScript Language Features`, then right click and select `Disable (Workspace)`. By default, Take Over mode will enable itself if the default TypeScript extension is disabled.
+2. Reload the VS Code window by running `Developer: Reload Window` from the command palette.
 
-  > alert-\<ID>
-
-3. 變數展示
-
-  > value:\<name>:\<value>
-
-4. 感測器數值展示
-
-  > sen:\<ADC>-\<BIT>
-
-## 範例
+You can learn more about Take Over mode [here](https://github.com/johnsoncodehk/volar/discussions/471).
 
 ```cpp
 #include <Arduino.h>
@@ -98,6 +89,3 @@ void debug_sensors() {
   Serial.println();
 }
 ```
-
-![alt text](.github/assets/home.png)
-![alt text](.github/assets/setting-1.png)
