--- extends/settings.json
@@ -1,60 +1,0 @@
-{
-  "alertSounds": {
-    "be": {
-      "name": "be",
-      "path": "alert-sounds/be.wav",
-      "description": "be",
-      "volume": 0.75,
-      "disable": false
-    },
-    "bee": {
-      "name": "bee",
-      "path": "alert-sounds/bee.wav",
-      "description": "bee",
-      "volume": 0.75,
-      "disable": false
-    },
-    "google-left": {
-      "name": "google-left",
-      "path": "alert-sounds/google-left.wav",
-      "description": "google-左",
-      "volume": 1,
-      "disable": false
-    },
-    "google-right": {
-      "name": "google-right",
-      "path": "alert-sounds/google-right.wav",
-      "description": "google-右",
-      "volume": 1,
-      "disable": false
-    },
-    "klee-bounce-bomb": {
-      "name": "klee-bounce-bomb",
-      "path": "alert-sounds/klee-bounce-bomb.mp3",
-      "description": "可莉-蹦蹦炸彈",
-      "volume": 1,
-      "disable": false
-    },
-    "klee-left": {
-      "name": "klee-left",
-      "path": "alert-sounds/klee-left.wav",
-      "description": "可莉-左",
-      "volume": 1,
-      "disable": false
-    },
-    "klee-pain": {
-      "name": "klee-pain",
-      "path": "alert-sounds/klee-pain.mp3",
-      "description": "可莉-好疼啊",
-      "volume": 1,
-      "disable": false
-    },
-    "klee-right": {
-      "name": "klee-right",
-      "path": "alert-sounds/klee-right.wav",
-      "description": "可莉-右",
-      "volume": 1,
-      "disable": false
-    }
-  }
-}
