+++ package.json
@@ -13,12 +13,12 @@
     "@primevue/themes": "^4.0.0",
     "@tauri-apps/api": "^1",
     "@types/node": "^20.14.11",
-    "@vueuse/components": "^10.11.0",
     "@vueuse/core": "^10.11.0",
     "pinia": "^2.1.7",
     "primeicons": "^7.0.0",
     "primevue": "^4.0.0",
-    "vue": "^3.3.4"
+    "vue": "^3.3.4",
+    "vue-router": "4"
   },
   "devDependencies": {
     "@tauri-apps/cli": "^1",
