+++ src-tauri/Cargo.toml
@@ -11,7 +11,7 @@ tauri-build = { version = "1", features = [] }
 
 [dependencies]
 serde = { version = "1", features = ["derive"] }
-tauri = { version = "1", features = ["fs-all", "shell-open", "devtools"] }
+tauri = { version = "1", features = ["shell-open", "devtools"] }
 serialport = { version = "4.4.0", features = ["serde"] }
 serde_json = "1"
 async-std = "1.12.0"
