+++ src/main.ts
@@ -1,5 +1,6 @@
 import { createApp } from 'vue';
 import { createPinia } from 'pinia';
+import { createMemoryHistory, createRouter } from 'vue-router';
 import PrimeVue from 'primevue/config';
 
 import Aura from '@primevue/themes/aura';
@@ -9,20 +10,23 @@ import 'primeicons/primeicons.css';
 
 import './style.css';
 import App from './App.vue';
-import Ripple from 'primevue/ripple';
 
 const pinia = createPinia();
 
-createApp(App)
-  .use(PrimeVue, {
-    theme: {
-      preset: Aura,
-      options: {
-        darkModeSelector: '.dark',
-      },
+const router = createRouter({
+  history: createMemoryHistory(),
+  routes: [
+    {
+      path: '/',
+      name: 'home',
+      component: () => import('./views/Home.vue'),
     },
-  })
-  .directive('ripple', Ripple) // fix for vue-warn
+  ],
+});
+
+createApp(App)
+  .use(PrimeVue, { theme: { preset: Aura } })
   .use(ToastService)
   .use(pinia)
+  .use(router)
   .mount('#app');
