+++ src/utils/index.ts
@@ -1,4 +1,2 @@
-export * from './event';
-export * from './format';
-export * from './serial-vue';
 export * from './serial';
+export * from './format';
