+++ tsconfig.json
@@ -1,7 +1,5 @@
 {
   "compilerOptions": {
-    "baseUrl": "./",
-
     "target": "ES2020",
     "useDefineForClassFields": true,
     "module": "ESNext",
@@ -20,11 +18,7 @@
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true,
-
-    "paths": {
-      "@/*": ["src/*"]
-    }
+    "noFallthroughCasesInSwitch": true
   },
   "include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.tsx", "src/**/*.vue"],
   "references": [{ "path": "./tsconfig.node.json" }]
